refactor(studio): extract useColorGradingController from ColorGradingSection
